#660 - store - split wishlist module
