+++ core/store/modules/wishlist/getters.js
@@ -1,0 +1,10 @@
+export default {
+  check (state) {
+    return {
+      isOnWishlist: (product) => {
+        let item = state.itemsWishlist.find(p => p.sku === product.sku)
+        return (item !== undefined)
+      }
+    }
+  }
+}
