+++ core/store/modules/wishlist/index.js
@@ -1,0 +1,14 @@
+import actions from './actions'
+import getters from './getters'
+import mutations from './mutations'
+
+export default {
+  namespaced: true,
+  state: {
+    wishlist: false,
+    itemsWishlist: []
+  },
+  getters,
+  actions,
+  mutations
+}
